requirements.txt
Flask
Flask-CORS
google-generativeai
python-dotenv
gunicorn